feature(orchestra.cpp): Now creates emtpy scenes
This is so important when a performance do not have any one. We could
choose to automatically add a "default" scene when a new performance is
created, but we prefered to mantain the program more "opened" and the
scene is created when the user tryes to open the ORCHESTRATION window.

The new method is called belongs to Orchestra class and it's called:
`add_empty_scene(const Performance*&)`. Now, the method is called
inmediatly the `ADD_VALUE` subroutine is called. In fact, it solves the
issue that occured when attempting to open `Orchestra` window if the
edited `Performance` has no scene.

Fixes #6