Fix issue when deleting the last element
The index was not moved when the last element was deleted, so it remains
"flying" and causing two known errors, the first is that no element was
finally selected; the second and fatal was that `print_zoom()` was
trying to print the value of a non actualized `dIndex` value.

The solution was simply force to move the index to the previous when the
last element is deleted.

Fixes #10